This is a PR for perf-remove-optional-creation.

Basically, we just need `std::string_view::begins_with` but that's C++20. However, we can roll our own.

Performance...

```
This PR:
BasicBench_AdaURL       5304 ns         5304 ns       131773 time/byte=6.56434ns time/url=482.181ns url/s=2.07391M/s
perf remove optional creation:
BasicBench_AdaURL       5373 ns         5372 ns       129493 time/byte=6.64898ns time/url=488.398ns url/s=2.04751M/s
Main branch:
BasicBench_AdaURL       5805 ns         5804 ns       120171 time/byte=7.18339ns time/url=527.653ns url/s=1.89519M/s
```